Move initializer to internal package
