+++ OptiGUI/src/main/kotlin/opekope2/optigui/internal/ResourceLoader.kt
@@ -7,7 +7,6 @@ import opekope2.optigui.interaction.Interaction
 import opekope2.optigui.internal.filter.IdentifiableFilter
 import opekope2.optigui.internal.interaction.filterFactories
 import opekope2.optigui.internal.service.ResourceLoaderService
-import opekope2.optigui.logger
 import opekope2.optigui.resource.Resource
 import opekope2.util.dump
